Classify SequentialBehaviour as CompoundBehaviour
[Changed]
- The OneShotBehaviour and CyclicBehaviour now inherits from SimpleBehaviour;
- A CompoundBehaviour class was created;
- The SequentialBehaviour class now inherits from CompoundBehaviour class;
- Updating the propoer doc;

[Fixed]
- Some spelling erros was fixed in doc;
