+++ pade/behaviours/base.py
@@ -33,7 +33,7 @@ the PADE Scheduler to manage the agent behaviours.
 
 from pade.behaviours.protocols import Behaviour
 from pade.acl.messages import ACLMessage
-import queue, time, threading
+import queue, time
 
 class BaseBehaviour(Behaviour):
 	''' The basic behaviour class.
@@ -45,12 +45,6 @@ class BaseBehaviour(Behaviour):
 	----------
 	_messages : list
 		The behaviour local message queue.
-	_lock : threding.Lock()
-		The lock object used to implement mutual exclusion.
-	_return : object
-		The data to be returned by this behaviour to another behaviour.
-	_event : threading.Event()
-		The event object used to implement the behaviour returning.
 	'''
 
 	def __init__(self, agent):
@@ -62,11 +56,7 @@ class BaseBehaviour(Behaviour):
 		'''
 
 		super().__init__(agent)
-		# Queue of received messages by the agent and unread by this behaviour
 		self._messages = queue.Queue()
-		self._lock = None
-		self._return = None
-		self._event = threading.Event()
 
 
 	def read(self, block = True):
@@ -211,75 +201,3 @@ class BaseBehaviour(Behaviour):
 		'''
 
 		return self._messages.qsize() != 0
-
-
-	def add_lock(self, lock):
-		''' Adds a threading.Lock object to this behaviour.
-
-		This allows the behaviour to execute the mutual exclusion.
-
-		Parameters
-		----------
-		lock : threadong.Lock
-			The lock object.
-		'''
-
-		self._lock = lock
-
-
-	@property	
-	def lock(self):
-		''' Returns the added lock object.
-
-		Raises
-		------
-		AttributeError
-			If there is no a lock object added.
-
-		Returns
-		-------
-		threading.Lock
-			The local lock object.
-		'''
-
-		if self._lock != None:
-			return self._lock
-		else:
-			raise(AttributeError('No such lock object added to this behaviour.'))
-
-
-	def set_return(self, data):
-		''' Sets the return for this behaviour.
-
-		Parameters
-		----------
-		data : object
-			The data to be returned by the behaviour.
-		'''
-
-		self._return = data
-		self._event.set()
-
-
-	def wait_return(self, behaviour, timeout = None):
-		''' Waits for the return from other behaviour.
-
-		Whether the target behaviour return is not set, this method
-		will block the behaviour until the return is set.
-
-		Parameters
-		----------
-		behaviour : BaseBehaviour
-			The behaviour you want to get the return.
-		timeout : float, optional
-			The max timeout to wait the target behaviour returns.
-
-		Returns
-		-------
-		object
-			The return data.
-		'''
-
-		behaviour._event.wait(timeout)
-		behaviour._event.clear()
-		return behaviour._return
