+++ pade/behaviours/types.py
@@ -33,6 +33,7 @@ used to model the agent behaviours of various types.
 
 from pade.behaviours.base import BaseBehaviour
 from pade.acl.messages import ACLMessage
+import queue, threading
 
 
 class SimpleBehaviour(BaseBehaviour):
@@ -41,7 +42,30 @@ class SimpleBehaviour(BaseBehaviour):
 	SimpleBehaviour class models a basic behaviour. The action() method must be
 	overridden in the subclasses to define the behaviour actions. The done()
 	method must indicate (by returning True) when this behaviour will finalize.
+
+	Attributes
+	----------
+	_lock : threding.Lock()
+		The lock object used to implement mutual exclusion.
+	_return : object
+		The data to be returned by this behaviour to another behaviour.
+	_event : threading.Event()
+		The event object used to implement the behaviour returning.
 	'''
+
+	def __init__(self, agent):
+		'''
+		Parameters
+		----------
+		agent : Agent
+			The agent which performs the behaviour.
+		'''
+
+		super().__init__(agent)
+		self._lock = None
+		self._return = None
+		self._event = threading.Event()
+
 
 	def action(self):
 		'''  An abstract method that performs the actions of the behaviour.
@@ -62,8 +86,153 @@ class SimpleBehaviour(BaseBehaviour):
 		pass
 
 
+	def add_lock(self, lock):
+		''' Adds a threading.Lock object to this behaviour.
 
-class OneShotBehaviour(BaseBehaviour):
+		This allows the behaviour to execute the mutual exclusion.
+
+		Parameters
+		----------
+		lock : threadong.Lock
+			The lock object.
+		'''
+
+		self._lock = lock
+
+
+	@property
+	def lock(self):
+		''' Returns the added lock object.
+
+		Raises
+		------
+		AttributeError
+			If there is no a lock object added.
+
+		Returns
+		-------
+		threading.Lock
+			The local lock object.
+		'''
+
+		if self._lock != None:
+			return self._lock
+		else:
+			raise(AttributeError('No such lock object added to this behaviour.'))
+
+
+	def set_return(self, data):
+		''' Sets the return for this behaviour.
+
+		Parameters
+		----------
+		data : object
+			The data to be returned by the behaviour.
+		'''
+
+		self._return = data
+		self._event.set()
+
+
+	def wait_return(self, behaviour, timeout = None):
+		''' Waits for the return from other behaviour.
+
+		Whether the target behaviour return is not set, this method
+		will block the behaviour until the return is set.
+
+		Parameters
+		----------
+		behaviour : BaseBehaviour
+			The behaviour you want to get the return.
+		timeout : float, optional
+			The max timeout to wait the target behaviour returns.
+
+		Returns
+		-------
+		object
+			The return data.
+		'''
+
+		behaviour._event.wait(timeout)
+		behaviour._event.clear()
+		return behaviour._return
+
+
+
+class CompoundBehaviour(BaseBehaviour):
+	''' This class models compound behaviours in PADE.
+
+	CompoundBehaviour is an abstract class to model behaviours that handle
+	other behaviours as sub-behaviours. This classe implements the general
+	add_subbehaviour(SimpleBehaviour) method that add simple behaviours in the
+	local queue as sub-behaviours. The actions of this class need to be
+	implemented in subclasses.
+
+	Attributes
+	----------
+	_subbehaviours : list
+		The list of added subbehaviours.
+	'''
+
+	def __init__(self, agent):
+		'''
+		Parameters
+		----------
+		agent : Agent
+			The agent that holds the behaviour.
+		'''
+
+		super().__init__(agent)
+		self.subbehaviours = list()
+
+
+	def add_subbehaviour(self, behaviour):
+		''' Adds sub-behaviours in this sub-behaviour local list.
+
+		Parameters
+		----------
+		behaviour : SimpleBehaviour
+			The simple behaviour to be added as sub-behaviour.
+
+		Raises
+		------
+		ValueError
+			If the sub-behaviour not it a subclass from SimpleBehaviour.
+		'''
+
+		if isinstance(behaviour, SimpleBehaviour):
+			self.subbehaviours.append(behaviour)
+		else:
+			raise ValueError('sub-behaviours must be of the type SimpleBehaviour!')
+
+
+	def receive(self, message):
+		''' Passes the received message to the sub-behaviours.
+
+		This method doesn't save the messages in the local queue, by default.
+		If you want to change that behaviour, override this method in the
+		subclasses.
+
+		Parameters
+		----------
+		message : ACLMessage
+			The message to be passed to the sub-behaviours.
+
+		Raises
+		------
+		ValueError
+			If the passed object is not an ACLMessage.
+		'''
+
+		if isinstance(message, ACLMessage):
+			for behaviour in self.subbehaviours:
+				behaviour.receive(message)
+		else:
+			raise ValueError('message object type must be ACLMessage!')
+
+
+
+class OneShotBehaviour(SimpleBehaviour):
 	''' This class models a finite behaviour.
 
 	OneShotBehaviour are behaviours that executes its action() method only 
@@ -90,7 +259,7 @@ class OneShotBehaviour(BaseBehaviour):
 
 
 
-class CyclicBehaviour(BaseBehaviour):
+class CyclicBehaviour(SimpleBehaviour):
 	''' This class models an infinite behaviour.
 
 	CyclicBehaviour are behaviours that executes its action() method until the
@@ -220,32 +389,14 @@ class TickerBehaviour(CyclicBehaviour):
 
 
 
-class SequentialBehaviour(OneShotBehaviour):
+class SequentialBehaviour(CompoundBehaviour):
 	''' This class models the sequential-compound behaviours in PADE.
 
 	SequentialBehaviour models a sequential-compound behaviour. This classe
-	adds other non-compound behaviours (like OneShotBehaviour or 
-	SimpleBehaviour) as sub-behaviours, and performs them sequentially. The
-	execution order is defined as the sub-behavious are added. No overrides are
-	required for this class.
-
-	Attributes
-	----------
-	_subbehaviours : list
-		The list of added subbehaviours.
+	adds SimpleBehaviour and its subclasses as sub-behaviours, and execute them
+	sequentially. The execution order is defined as the sub-behavious are
+	added. No overrides are required for this class.
 	'''
-
-	def __init__(self, agent):
-		'''
-		Parameters
-		----------
-		agent : Agent
-			The agent that holds the behaviour.
-		'''
-
-		super().__init__(agent)
-		self.subbehaviours = list()
-
 
 	def action(self):
 		''' This method performs the actions of the behaviour.
@@ -261,34 +412,11 @@ class SequentialBehaviour(OneShotBehaviour):
 			behaviour.on_end()
 
 
-	def add_subbehaviour(self, behaviour):
-		''' Adds sub-behaviours in this sub-behaviour local list.
+	def done(self):
+		''' Defines when the behaviour ends.
 
-		Parameters
-		----------
-		behaviour : BaseBehaviour
-			The behaviour to be added as sub-behaviour.
+		This method always returns True and should not be overridden in the
+		subclasses. By returning True, the behaviour will execute only once.
 		'''
 
-		self.subbehaviours.append(behaviour)
-
-
-	def receive(self, message):
-		''' Passes the received message to the sub-behaviours.
-
-		Parameters
-		----------
-		message : ACLMessage
-			The message to be passed to the sub-behaviours.
-
-		Raises
-		------
-		ValueError
-			If the passed object is not an ACLMessage.
-		'''
-
-		if isinstance(message, ACLMessage):
-			for behaviour in self.subbehaviours:
-				behaviour.receive(message)
-		else:
-			raise ValueError('message object type must be ACLMessage!')
+		return True
